requirements.txt
google-generativeai
pymupdf
ipywidgets
requests
